Set InstructionSchema as a pydantic model
