+++ llm_data_cleaner/cleaner.py
@@ -2,9 +2,25 @@ import os
 import pandas as pd
 from typing import Dict, Any, Type, List, Optional
 from openai import OpenAI
-from pydantic import BaseModel, create_model
+from pydantic import BaseModel, RootModel, create_model
 import time
 from tqdm import tqdm
+
+# schema for instructions as a pydantic model
+# instructions is a dictionary, with keys as column names and values as dictionaries of "prompt" (a string) amd "schema" (a pydantic model)
+class InstructionField(BaseModel):
+    prompt: str
+    schema: Type[BaseModel]  # Pydantic model class for the schema
+
+class InstructionSchema(RootModel):
+    root: Dict[str, InstructionField]
+
+    def __iter__(self):
+        return iter(self.root)
+
+    def __getitem__(self, item):
+        return self.root[item]
+
 
 class DataCleaner:
     """
@@ -38,7 +54,7 @@ class DataCleaner:
     def clean_dataframe(
         self,
         df: pd.DataFrame,
-        instructions: Dict[str, Dict[str, Any]],
+        instructions: InstructionSchema,
     ) -> pd.DataFrame:
         """
         Cleans each column specified in instructions, in batches.
@@ -143,54 +159,3 @@ class DataCleaner:
             batch_name,
             cleaned=(List[Optional[model]], ...)
         )
-   
-    @staticmethod
-    def load_yaml_models(yaml_path:str = None) -> Dict[str, Type[BaseModel]]:
-        """
-        Load models from YAML file.
-        """
-        import yaml
-        def parse_type(props, model_defs=None, model_cache=None):
-
-            if model_defs is None:
-                model_defs = {}
-            if model_cache is None:
-                model_cache = {}
-
-            typ = props["type"]
-            optional = props.get("optional", False)
-            if typ == 'str':
-                typ = str
-
-            if typ == 'float':
-                typ = float
-
-            if typ == 'list':
-                items = props["items"]
-                if isinstance(items, dict):
-                    typ = List[parse_type(items, model_defs, model_cache)]
-                else:
-                    typ = List[parse_type({"type": items}, model_defs, model_cache)]
-
-            if optional:
-                typ = Optional[typ]
-
-            return typ
-
-        with open(yaml_path, "r") as f:
-            schema = yaml.safe_load(f)
-        models = {}
-        for name, model_def in schema.items():
-            fields = model_def["fields"]
-            annotations = {}
-            defaults = {}
-            for field, props in fields.items():
-                t = parse_type(props)
-                print(t)
-                if props.get("optional", False):
-                    defaults[field] = None
-                    annotations[field] = (t, None)
-                else:
-                    annotations[field] = (t, ...)
-            models[name] = create_model(name, **annotations, __base__=BaseModel)
-        return models
